pyproject.toml
[tool.poetry]
name = "project-tracker"
version = "0.1.0"
description = "Project Tracker backend service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.116.1"}
sqlalchemy = "^2.0.42"
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
asyncpg = "^0.30.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
wemake-python-styleguide = "^1.3.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
